mcp/gateway/requirements.txt
fastmcp>=2.3.0
pyyaml>=6.0
starlette>=0.37
uvicorn>=0.30
